backend/requirements.txt
fastapi
uvicorn
python-dotenv
psycopg[binary]
langchain-community
langchain-text-splitters
sentence-transformers
FlagEmbedding
torch

# Step 3.1 dependencies
pandas>=2.0.0
openpyxl>=3.1.0
pm4py>=2.7.0
pytest>=7.4.0

# Step 3.2 & 4 dependencies (LangGraph agent)
langgraph>=0.0.20
langchain-openai>=0.0.5
langchain-core>=0.1.0
